Get further in the build when using Clang trunk
Clang defaults to C99 mode, which doesn't permit extern inline
re-definitions such as those found in the gcc source. For now,
we check to see if we're using Clang, and if so, add the flag
to choose gnu89 mode. Ideally, we'd test to see if the compiler
accepted extern inline re-definitions, and if not, we'd try one
or more flags to resolve it. This would also allow lack of such
flags on other compilers to be caught at configure time, rather
than 5% into the entire compile of a full suite.